setup cmd line usage
