+++ Cargo.toml
@@ -2,14 +2,23 @@
 edition = "2021"
 name = "soe-network-parser"
 version = "0.1.0"
-
 [dependencies]
-h1emu-core = { version = "0.8.9", default-features = false, features = ["soeprotocol"] }
+hex = "0.4.3"
+path-absolutize = "3.0.13"
 serde = "1.0.136"
 serde_derive = "1.0.136"
-serde_json ="1.0.79"
-hex = "0.4.3"
+serde_json = "1.0.79"
 
+[dependencies.clap]
+features = ["derive"]
+version = "3.1.18"
+
+[dependencies.h1emu-core]
+default-features = false
+features = ["soeprotocol"]
+version = "0.8.9"
+
+[profile]
 [profile.release]
 lto = true
 opt-level = 3
